+++ docs/changes.rst
@@ -30,7 +30,7 @@ All traces of Grok have been removed.
 6.2.22 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Update Statistics: Order results to keep them stable during batching.
 
 
 6.2.21 (2021-04-28)